Safari-611.1.10/LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-valid-expected.txt


FAIL [INPUT in TEXT status] validity.valid must be false if validity.tooLong is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in TEXT status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in TEXT status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in SEARCH status] validity.valid must be false if validity.tooLong is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in SEARCH status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in SEARCH status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in TEL status] validity.valid must be false if validity.tooLong is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in TEL status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in TEL status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in PASSWORD status] validity.valid must be false if validity.tooLong is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in PASSWORD status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in PASSWORD status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in URL status] validity.valid must be false if validity.tooLong is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in URL status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in URL status] validity.valid must be false if validity.typeMismatch is true
PASS [INPUT in URL status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in EMAIL status] validity.valid must be false if validity.tooLong is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in EMAIL status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in EMAIL status] validity.valid must be false if validity.typeMismatch is true
PASS [INPUT in EMAIL status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in DATETIME status] The datetime type must be supported. assert_equals: The datetime type should be supported. expected "datetime" but got "text"
PASS [INPUT in DATE status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in DATE status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in DATE status] validity.valid must be false if validity.stepMismatch is true
PASS [INPUT in DATE status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.rangeUnderflow is true
FAIL [INPUT in MONTH status] validity.valid must be false if validity.stepMismatch is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in MONTH status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.stepMismatch is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in TIME status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in TIME status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in TIME status] validity.valid must be false if validity.stepMismatch is true
PASS [INPUT in TIME status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in NUMBER status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in NUMBER status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in NUMBER status] validity.valid must be false if validity.stepMismatch is true
PASS [INPUT in NUMBER status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in CHECKBOX status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in RADIO status] validity.valid must be false if validity.valueMissing is true
PASS [INPUT in FILE status] validity.valid must be false if validity.valueMissing is true
PASS [select]  validity.valid must be false if validity.valueMissing is true
PASS [textarea]  validity.valid must be false if validity.valueMissing is true
